Retain error event details
