+++ tests/unit/tasks/test_exceptions.py
